requirements.txt
tqdm
rotary_embedding_torch
transformers==4.29.2
tokenizers
inflect
progressbar
einops==0.4.1
unidecode
scipy
librosa==0.9.1
ffmpeg
IPython
numpy
numba
torchaudio
threadpoolctl
llvmlite
appdirs
nbconvert==5.3.1
tornado==4.2
pydantic==1.9.1
deepspeed==0.8.3
py-cpuinfo
hjson
psutil
sounddevice
spacy==3.7.5
